Refactored the template redirect functions into a class.
